test_fields (wip)
