+++ odoo/addons/test_orm/tests/__init__.py
@@ -16,6 +16,15 @@ from . import (
     test_domain,
     test_domain_expression,
     test_fields,
+    test_fields_binary,
+    test_fields_misc,
+    test_fields_numeric,
+    test_fields_properties,
+    test_fields_reference,
+    test_fields_relational,
+    test_fields_selection,
+    test_fields_temporal,
+    test_fields_textual,
     test_html_converter,
     test_indexed_translation,
     test_inherits,
