+++ odoo/addons/test_orm/tests/test_fields.py
@@ -633,13 +633,6 @@ class TestFields(TransactionCaseWithUserDemo, TransactionExpressionCase):
         self.registry._setup_models__(self.cr, ['test_orm.compute.dynamic.depends'])
         self.assertEqual(self.registry.field_depends[Model.full_name], ('name1', 'name2'))
 
-    def test_12_one2many_reference_domain(self):
-        model = self.env['test_orm.inverse_m2o_ref']
-        o2m_field = model._fields['model_ids']
-        self.assertEqual(o2m_field.get_comodel_domain(model), Domain('const', '=', True) & Domain('res_model', '=', model._name))
-        o2m_field = model._fields['model_computed_ids']
-        self.assertEqual(o2m_field.get_comodel_domain(model), Domain.TRUE)
-
     def test_13_inverse(self):
         """ test inverse computation of fields """
         Category = self.env['test_orm.category']
@@ -1017,109 +1010,6 @@ class TestFields(TransactionCaseWithUserDemo, TransactionExpressionCase):
         self.assertEqual(count, 0, "Precommit not ran enough times")
         self.assertEqual(record.value, 10, "Flush not triggered correctly")
 
-    def test_20_float(self):
-        """ test rounding of float fields """
-        record = self.env['test_orm.mixed'].create({})
-        query = "SELECT 1 FROM test_orm_mixed WHERE id=%s AND number=%s"
-
-        # 2.49609375 (exact float) must be rounded to 2.5
-        record.write({'number': 2.49609375})
-        self.env.flush_all()
-        self.cr.execute(query, [record.id, '2.5'])
-        self.assertTrue(self.cr.rowcount)
-        self.assertEqual(record.number, 2.5)
-
-        # 1.1 (1.1000000000000000888178420 in float) must be 1.1 in database
-        record.write({'number': 1.1})
-        self.env.flush_all()
-        self.cr.execute(query, [record.id, '1.1'])
-        self.assertTrue(self.cr.rowcount)
-        self.assertEqual(record.number, 1.1)
-
-    def test_21_float_digits(self):
-        """ test field description """
-        precision = self.env.ref('test_orm.decimal_orm_number')
-        description = self.env['test_orm.mixed'].fields_get()['number2']
-        self.assertEqual(description['digits'], (16, precision.digits))
-
-    def check_monetary(self, record, amount, currency, msg=None):
-        # determine the possible roundings of amount
-        if currency:
-            ramount = currency.round(amount)
-            samount = float(float_repr(ramount, currency.decimal_places))
-        else:
-            ramount = samount = amount
-
-        # check the currency on record
-        self.assertEqual(record.currency_id, currency)
-
-        # check the value on the record
-        self.assertIn(record.amount, [ramount, samount], msg)
-
-        # check the value in the database
-        self.env.flush_all()
-        self.cr.execute('SELECT amount FROM test_orm_mixed WHERE id=%s', [record.id])
-        value = self.cr.fetchone()[0]
-        self.assertEqual(value, samount, msg)
-
-    def test_20_monetary(self):
-        """ test monetary fields """
-        model = self.env['test_orm.mixed']
-        currency = self.env['res.currency'].with_context(active_test=False)
-        amount = 14.70126
-
-        for rounding in [0.01, 0.0001, 1.0, 0]:
-            # first retrieve a currency corresponding to rounding
-            if rounding:
-                currency = currency.search([('rounding', '=', rounding)], limit=1)
-                self.assertTrue(currency, "No currency found for rounding %s" % rounding)
-            else:
-                # rounding=0 corresponds to currency=False
-                currency = currency.browse()
-
-            # case 1: create with amount and currency
-            record = model.create({'amount': amount, 'currency_id': currency.id})
-            self.check_monetary(record, amount, currency, 'create(amount, currency)')
-
-            # case 2: assign amount
-            record.amount = 0
-            record.amount = amount
-            self.check_monetary(record, amount, currency, 'assign(amount)')
-
-            # case 3: write with amount and currency
-            record.write({'amount': 0, 'currency_id': False})
-            record.write({'amount': amount, 'currency_id': currency.id})
-            self.check_monetary(record, amount, currency, 'write(amount, currency)')
-
-            # case 4: write with amount only
-            record.write({'amount': 0})
-            record.write({'amount': amount})
-            self.check_monetary(record, amount, currency, 'write(amount)')
-
-            # case 5: write with amount on several records
-            records = record + model.create({'currency_id': currency.id})
-            records.write({'amount': 0})
-            records.write({'amount': amount})
-            for record in records:
-                self.check_monetary(record, amount, currency, 'multi write(amount)')
-
-    def test_20_monetary_related(self):
-        """ test value rounding with related currency """
-        currency = self.env.ref('base.USD')
-        monetary_base = self.env['test_orm.monetary_base'].create({
-            'base_currency_id': currency.id,
-        })
-        monetary_related = self.env['test_orm.monetary_related'].create({
-            'monetary_id': monetary_base.id,
-            'total': 1 / 3,
-        })
-        self.env.cr.execute(
-            "SELECT total FROM test_orm_monetary_related WHERE id=%s",
-            monetary_related.ids,
-        )
-        [total] = self.env.cr.fetchone()
-        self.assertEqual(total, .33)
-
     def test_20_like(self):
         """ test filtered_domain() on char fields. """
         record = self.env['test_orm.multi.tag'].create({'name': 'Foo'})
@@ -1139,115 +1029,6 @@ class TestFields(TransactionCaseWithUserDemo, TransactionExpressionCase):
         record = self.env['test_orm.mixed'].create({'comment1': 'Foo\nBar'})
         self.assertTrue(record.filtered_domain([('comment1', 'like', 'Bar')]))
         self.assertTrue(record.filtered_domain([('comment1', 'ilike', 'bar')]))
-
-    def test_21_date(self):
-        """ test date fields """
-        record = self.env['test_orm.mixed'].create({})
-
-        # one may assign False or None
-        record.date = None
-        self.assertFalse(record.date)
-
-        # one may assign date but not datetime objects
-        record.date = date(2012, 5, 1)
-        self.assertEqual(record.date, date(2012, 5, 1))
-
-        # DLE P41: We now support to assign datetime to date. Not sure this is the good practice though.
-        # with self.assertRaises(TypeError):
-        #     record.date = datetime(2012, 5, 1, 10, 45, 0)
-
-        # one may assign dates and datetime in the default format, and it must be checked
-        record.date = '2012-05-01'
-        self.assertEqual(record.date, date(2012, 5, 1))
-
-        record.date = "2012-05-01 10:45:00"
-        self.assertEqual(record.date, date(2012, 5, 1))
-
-        with self.assertRaises(ValueError):
-            record.date = '12-5-1'
-
-        # check filtered_domain
-        self.assertTrue(record.filtered_domain([('date', '<', '2012-05-02')]))
-        self.assertTrue(record.filtered_domain([('date', '<', date(2012, 5, 2))]))
-        self.assertTrue(record.filtered_domain([('date', '<', datetime(2012, 5, 2, 12, 0, 0))]))
-        self.assertTrue(record.filtered_domain([('date', '!=', False)]))
-        self.assertFalse(record.filtered_domain([('date', '=', False)]))
-
-        record.date = None
-        self.assertFalse(record.filtered_domain([('date', '<', '2012-05-02')]))
-        self.assertFalse(record.filtered_domain([('date', '<', date(2012, 5, 2))]))
-        self.assertFalse(record.filtered_domain([('date', '<', datetime(2012, 5, 2, 12, 0, 0))]))
-        self.assertFalse(record.filtered_domain([('date', '!=', False)]))
-        self.assertTrue(record.filtered_domain([('date', '=', False)]))
-
-    def test_21_datetime(self):
-        """ test datetime fields """
-        for _i in range(0, 10):
-            self.assertEqual(fields.Datetime.now().microsecond, 0)
-
-        record = self.env['test_orm.mixed'].create({})
-
-        # assign falsy value
-        record.moment = None
-        self.assertFalse(record.moment)
-
-        # assign string
-        record.moment = '2012-05-01'
-        self.assertEqual(record.moment, datetime(2012, 5, 1))
-        record.moment = '2012-05-01 06:00:00'
-        self.assertEqual(record.moment, datetime(2012, 5, 1, 6))
-        with self.assertRaises(ValueError):
-            record.moment = '12-5-1'
-
-        # assign date or datetime
-        record.moment = date(2012, 5, 1)
-        self.assertEqual(record.moment, datetime(2012, 5, 1))
-        record.moment = datetime(2012, 5, 1, 6)
-        self.assertEqual(record.moment, datetime(2012, 5, 1, 6))
-
-        # check filtered_domain
-        self.assertTrue(record.filtered_domain([('moment', '<', '2012-05-02')]))
-        self.assertTrue(record.filtered_domain([('moment', '<', date(2012, 5, 2))]))
-        self.assertTrue(record.filtered_domain([('moment', '<', datetime(2012, 5, 1, 12, 0, 0))]))
-        self.assertTrue(record.filtered_domain([('moment', '!=', False)]))
-        self.assertFalse(record.filtered_domain([('moment', '=', False)]))
-
-        record.moment = None
-        self.assertFalse(record.filtered_domain([('moment', '<', '2012-05-02')]))
-        self.assertFalse(record.filtered_domain([('moment', '<', date(2012, 5, 2))]))
-        self.assertFalse(record.filtered_domain([('moment', '<', datetime(2012, 5, 2, 12, 0, 0))]))
-        self.assertFalse(record.filtered_domain([('moment', '!=', False)]))
-        self.assertTrue(record.filtered_domain([('moment', '=', False)]))
-
-    def test_21_date_dynamic(self):
-        record = self.env['test_orm.mixed'].create({'moment': fields.Datetime.now()})
-        self.assertEqual(record, self._search(record, [('moment', '<', 'now +1d')], [('id', 'in', record.ids)]))
-        self.assertFalse(self._search(record, [('moment', '<', 'today')], [('id', 'in', record.ids)]))
-        self.assertEqual(record, self._search(record, [('moment', '>', '-1H')], [('id', 'in', record.ids)]))
-
-    def test_22_selection(self):
-        """ test selection fields """
-        record_list = self.env['test_orm.selection'].create({})
-        self.assertIsInstance(record_list._fields['state'].selection, list)
-
-        # the following selection is defined by a callable (method name)
-        record_call = self.env['test_orm.mixed'].create({})
-        self.assertIsInstance(record_call._fields['lang'].selection, str)
-
-        # one may assign a value
-        record_list.state = 'foo'
-        record_call.lang = self.env['res.lang'].search([], limit=1).code
-
-        # one may assign False or None
-        record_list.state = None
-        self.assertFalse(record_list.state)
-        record_call.lang = None
-        self.assertFalse(record_call.lang)
-
-        # the assigned value is only checked for the list case
-        with self.assertRaises(ValueError):
-            record_list.state = 'zz_ZZ'
-        record_call.lang = 'zz_ZZ'
 
     def test_23_relation(self):
         """ test relation fields """
@@ -1444,46 +1225,6 @@ class TestFields(TransactionCaseWithUserDemo, TransactionExpressionCase):
         bar.name = 'B'
         self.assertEqual(bar.foo, oof)
         self.assertIn(bar, bar.search([('foo', 'in', oof.ids)]))
-
-    def test_25_related_many2one(self):
-        bar = self.env['test_orm.related_bar'].create({'name': 'A'})
-        foo = self.env['test_orm.related_foo'].create({'name': 'A', 'bar_id': bar.id})
-        self.assertEqual(foo.bar_id, bar)
-        self.assertEqual(foo.bar_alias, foo.bar_id)
-
-        # After deactivating the foo record, the search should be executed with
-        # context depending on searching a many2one field: active_test=False.
-        for active in (True, False):
-            with self.subTest(active=active):
-                bar.active = active
-                self.assertEqual(foo.search([('id', 'in', foo.ids), ('bar_id', 'ilike', 'A')]), foo)
-                self.assertEqual(foo.search([('id', 'in', foo.ids), ('bar_alias', 'ilike', 'A')]), foo)
-
-    def test_25_one2many_inverse_related(self):
-        left = self.env['test_orm.trigger.left'].create({})
-        right = self.env['test_orm.trigger.right'].create({})
-        self.assertFalse(left.right_id)
-        self.assertFalse(right.left_ids)
-        self.assertFalse(right.left_size)
-
-        # create middle: this should trigger left.right_id by traversing
-        # middle.left_id, and right.left_size by traversing left.right_id
-        # after its computation!
-        middle = self.env['test_orm.trigger.middle'].create({
-            'left_id': left.id,
-            'right_id': right.id,
-        })
-        self.assertEqual(left.right_id, right)
-        self.assertEqual(right.left_ids, left)
-        self.assertEqual(right.left_size, 1)
-
-        # delete middle: this should trigger left.right_id by traversing
-        # middle.left_id, and right.left_size by traversing left.right_id
-        # before its computation!
-        middle.unlink()
-        self.assertFalse(left.right_id)
-        self.assertFalse(right.left_ids)
-        self.assertFalse(right.left_size)
 
     def test_26_inherited(self):
         """ test inherited fields. """
@@ -2325,58 +2066,6 @@ class TestFields(TransactionCaseWithUserDemo, TransactionExpressionCase):
         self.assertEqual(new_move.quantity, 3)
         self.assertEqual(move.quantity, 2)
 
-    def test_41_new_one2many(self):
-        """ Check command on one2many field on new record. """
-        move = self.env['test_orm.move'].create({})
-        line = self.env['test_orm.move_line'].create({'move_id': move.id, 'quantity': 1})
-        self.env.flush_all()
-
-        new_move = move.new(origin=move)
-        new_line = line.new(origin=line)
-        self.assertEqual(new_move.line_ids, new_line)
-
-        # drop line, and create a new one
-        new_move.line_ids = [Command.delete(new_line.id), Command.create({'quantity': 2})]
-        self.assertEqual(len(new_move.line_ids), 1)
-        self.assertFalse(new_move.line_ids.id)
-        self.assertEqual(new_move.line_ids.quantity, 2)
-
-        # assign line to new move without origin
-        new_move = move.new()
-        new_move.line_ids = line
-        self.assertFalse(new_move.line_ids.id)
-        self.assertEqual(new_move.line_ids._origin, line)
-        self.assertEqual(new_move.line_ids.move_id, new_move)
-
-    def test_41_new_many2many(self):
-        group = self.env['test_orm.group'].create({})
-        user0 = self.env['test_orm.user'].create({'group_ids': [Command.link(group.id)]})
-        new_user0 = user0.new(origin=user0)
-        new_group = group.new(origin=group)
-
-        self.env.invalidate_all()
-
-        # creating new_user1 shoud not fetch new_group.all_user_ids, which is the
-        # inverse of field new_user1.group_ids
-        with self.assertQueryCount(0):
-            new_user1 = self.env['test_orm.user'].new({'group_ids': [Command.link(group.id)]})
-            self.assertEqual(new_user1.group_ids, new_group)
-
-        # accessing new_group.all_user_ids should fetch group.all_user_ids and patch
-        # new_group.all_user_ids
-        with self.assertQueryCount(1):
-            self.assertEqual(new_group.user_ids, new_user0 + new_user1)
-
-        # creating new_user2 should patch new_group.all_user_ids immediately, since
-        # it is in cache
-        with self.assertQueryCount(0):
-            new_user2 = self.env['test_orm.user'].new({'group_ids': [Command.link(group.id)]})
-            self.assertEqual(new_user2.group_ids, new_group)
-            self.assertEqual(new_group.user_ids, new_user0 + new_user1 + new_user2)
-
-        # the patches on new_group.all_user_ids should not have changed group.all_user_ids
-        self.assertEqual(group.user_ids, user0)
-
     @mute_logger('odoo.addons.base.models.ir_model')
     def test_41_new_related(self):
         """ test the behavior of related fields starting on new records. """
@@ -2441,78 +2130,6 @@ class TestFields(TransactionCaseWithUserDemo, TransactionExpressionCase):
         defaults = self.env['test_orm.mixed'].default_get(['number'])
         self.assertEqual(defaults, {'number': 3.14})
 
-    def test_50_search_many2one(self):
-        """ test search through a path of computed fields"""
-        messages = self.env['test_orm.message'].search(
-            [('author_partner.name', '=', 'Marc Demo')])
-        self.assertEqual(messages, self.env.ref('test_orm.message_0_1'))
-
-    def test_51_search_many2one_ordered(self):
-        """ test search on many2one ordered by id """
-        with self.assertQueries(['''
-            SELECT "test_orm_message"."id" FROM "test_orm_message"
-            WHERE "test_orm_message"."active" IS TRUE
-            ORDER BY  "test_orm_message"."discussion"
-        ''']):
-            self.env['test_orm.message'].search([], order='discussion')
-
-    def test_52_search_many2one_active_test(self):
-        Model = self.env['test_orm.model_active_field']
-
-        active_parent = Model.create({'name': 'Parent'})
-        child_of_active = Model.create({'parent_id': active_parent.id})
-
-        inactive_parent = Model.create({'name': 'Parent', 'active': False})
-        child_of_inactive = Model.create({'parent_id': inactive_parent.id})
-
-        self.assertEqual(
-            self._search(Model, [('parent_id.name', '=', 'Parent')]),
-            child_of_active + child_of_inactive,
-        )
-        self.assertEqual(
-            self._search(Model, [('parent_id', '=', 'Parent')]),
-            child_of_active + child_of_inactive,
-        )
-        self.assertEqual(
-            Model.search([('id', 'child_of', active_parent.id)]),
-            active_parent + child_of_active,
-        )
-        # weird semantics: active_parent is in both results but doesn't have a parent_id
-        self.assertEqual(
-            self._search(Model, [('parent_id', 'child_of', active_parent.id)]),
-            active_parent + child_of_active,
-        )
-        self.assertEqual(
-            self._search(Model, [('parent_id', 'child_of', 'Parent')]),
-            active_parent + child_of_active + child_of_inactive,
-        )
-
-    def test_53_boolean_query(self):
-        Model = self.env['test_orm.model_active_field']
-
-        with self.assertQueries(["""
-            SELECT "test_orm_model_active_field"."id"
-            FROM "test_orm_model_active_field"
-            WHERE "test_orm_model_active_field"."active" IS TRUE
-            ORDER BY "test_orm_model_active_field"."id"
-        """, """
-            SELECT "test_orm_model_active_field"."id"
-            FROM "test_orm_model_active_field"
-            WHERE "test_orm_model_active_field"."active" IS NOT TRUE
-            ORDER BY "test_orm_model_active_field"."id"
-        """]):
-            Model.search([('active', '=', True)])
-            Model.search([('active', '=', False)])
-
-        with self.assertQueries(["""
-            SELECT "test_orm_model_active_field"."id"
-            FROM "test_orm_model_active_field"
-            ORDER BY "test_orm_model_active_field"."id"
-        """]):
-            Model.search([('active', 'in', [True, False])])
-        with self.assertQueries([]):
-            Model.search([('active', 'not in', [True, False])])
-
     def test_54_not_null_id_query(self):
         # Test at post_install since not_null_fields is only loaded at the end of the registry
         Model = self.env['test_orm.model_active_field'].with_context(active_test=False)
@@ -2527,164 +2144,6 @@ class TestFields(TransactionCaseWithUserDemo, TransactionExpressionCase):
         """]):
             Model.search([('id', '!=', 1)])
             Model.search([('id', '=', False)])  # No query
-
-    def test_60_one2many_domain(self):
-        """ test the cache consistency of a one2many field with a domain """
-        discussion = self.env.ref('test_orm.discussion_0')
-        message = discussion.messages[0]
-        self.assertNotIn(message, discussion.important_messages)
-
-        message.important = True
-        self.assertIn(message, discussion.important_messages)
-
-        # writing on very_important_messages should call its domain method
-        self.assertIn(message, discussion.very_important_messages)
-        discussion.write({'very_important_messages': [Command.clear()]})
-        self.assertFalse(discussion.very_important_messages)
-        self.assertFalse(message.exists())
-
-    def test_60_many2many_domain(self):
-        """ test the cache consistency of a many2many field with a domain """
-        tag = self.env['test_orm.multi.tag'].create({'name': 'bar'})
-        record = self.env['test_orm.multi'].create({'tags': tag.ids})
-        self.env.flush_all()
-        self.env.invalidate_all()
-
-        self.assertEqual(type(record).tags.domain, [('name', 'ilike', 'a')])
-
-        # the tag is in the many2many
-        self.assertIn(tag, record.tags)
-
-        # modify the tag; it should not longer be in the many2many
-        tag.name = 'foo'
-        self.assertNotIn(tag, record.tags)
-
-        # modify again the tag; it should be back in the many2many
-        tag.name = 'baz'
-        self.assertIn(tag, record.tags)
-
-    def test_61_one2many_domain(self):
-        model = self.env['test_orm.inverse_m2o_ref']
-        field = model._fields['model_ids']
-        self.assertEqual(
-            field.get_comodel_domain(model),
-            Domain('const', '=', True) & Domain('res_model', '=', model._name),
-        )
-        self.assertEqual(
-            field.get_description(self.env, ['domain'])['domain'],
-            "([('const', '=', True)]) + ([('res_model', '=', 'test_orm.inverse_m2o_ref')])",
-            "res_model should appear in the descripton of the domain",
-        )
-
-    def test_70_x2many_write(self):
-        discussion = self.env.ref('test_orm.discussion_0')
-        # See YTI FIXME
-        self.env.invalidate_all()
-
-        Message = self.env['test_orm.message']
-        # There must be 3 messages, 0 important
-        self.assertEqual(len(discussion.messages), 3)
-        self.assertEqual(len(discussion.important_messages), 0)
-        self.assertEqual(len(discussion.very_important_messages), 0)
-        discussion.important_messages = [Command.create({
-            'body': 'What is the answer?',
-            'important': True,
-        })]
-        # There must be 4 messages, 1 important
-        self.assertEqual(len(discussion.messages), 4)
-        self.assertEqual(len(discussion.important_messages), 1)
-        self.assertEqual(len(discussion.very_important_messages), 1)
-        discussion.very_important_messages |= Message.new({
-            'body': '42',
-            'important': True,
-        })
-        # There must be 5 messages, 2 important
-        self.assertEqual(len(discussion.messages), 5)
-        self.assertEqual(len(discussion.important_messages), 2)
-        self.assertEqual(len(discussion.very_important_messages), 2)
-
-    def test_70_relational_inverse(self):
-        """ Check the consistency of relational fields with inverse(s). """
-        discussion = self.env.ref('test_orm.discussion_0')
-        demo_discussion = discussion.with_user(self.user_demo)
-
-        # check that the demo user sees the same messages
-        self.assertEqual(demo_discussion.messages, discussion.messages)
-
-        # See YTI FIXME
-        self.env.flush_all()
-        self.env.invalidate_all()
-
-        # add a message as user demo
-        messages = demo_discussion.messages
-        message = messages.create({'discussion': discussion.id})
-        self.assertEqual(demo_discussion.messages, messages + message)
-        self.assertEqual(demo_discussion.messages, discussion.messages)
-
-        # add a message as superuser
-        messages = discussion.messages
-        message = messages.create({'discussion': discussion.id})
-        self.assertEqual(discussion.messages, messages + message)
-        self.assertEqual(demo_discussion.messages, discussion.messages)
-
-    def test_71_relational_inverse(self):
-        """ Check the consistency of relational fields with inverse(s). """
-        move1 = self.env['test_orm.move'].create({})
-        move2 = self.env['test_orm.move'].create({})
-        line = self.env['test_orm.move_line'].create({'move_id': move1.id})
-        self.env.flush_all()
-        self.env.invalidate_all()
-
-        line.with_context(prefetch_fields=False).move_id
-
-        # Setting 'move_id' updates the one2many field that is based on it,
-        # which has a domain.  Here we check that evaluating the domain does not
-        # accidentally override 'move_id' (by prefetch).
-        line.move_id = move2
-        self.assertEqual(line.move_id, move2)
-
-    def test_72_relational_inverse(self):
-        """ Check the consistency of relational fields with inverse(s). """
-        move1 = self.env['test_orm.move'].create({})
-        move2 = self.env['test_orm.move'].create({})
-
-        # makes sure that line.move_id is flushed before search
-        line = self.env['test_orm.move_line'].create({'move_id': move1.id})
-        moves = self.env['test_orm.move'].search([('line_ids', 'in', line.id)])
-        self.assertEqual(moves, move1)
-
-        # makes sure that line.move_id is flushed before search
-        line.move_id = move2
-        moves = self.env['test_orm.move'].search([('line_ids', 'in', line.id)])
-        self.assertEqual(moves, move2)
-
-    def test_73_relational_inverse(self):
-        """ Check the consistency of relational fields with inverse(s). """
-        discussion1, discussion2 = self.env['test_orm.discussion'].create([
-            {'name': "discussion1"}, {'name': "discussion2"},
-        ])
-        category1, category2 = self.env['test_orm.category'].create([
-            {'name': "category1"}, {'name': "category2"},
-        ])
-
-        # assumption: category12 and category21 are in different order, but are
-        # in the same order when put in a set()
-        category12 = category1 + category2
-        category21 = category2 + category1
-        self.assertNotEqual(category12.ids, category21.ids)
-        self.assertEqual(list(set(category12.ids)), list(set(category21.ids)))
-
-        # make sure discussion1.categories is in cache; the write() below should
-        # update the cache of discussion1.categories by appending category12.ids
-        discussion1.categories
-        category12.write({'discussions': [Command.link(discussion1.id)]})
-        self.assertEqual(discussion1.categories.ids, category12.ids)
-
-        # make sure discussion2.categories is in cache; the write() below should
-        # update the cache of discussion2.categories by appending category21.ids
-        discussion2.categories
-        category21.write({'discussions': [Command.link(discussion2.id)]})
-        self.assertEqual(discussion2.categories.ids, category21.ids)
 
     def test_80_copy(self):
         discussion = self.env.ref('test_orm.discussion_0')
@@ -2710,334 +2169,6 @@ class TestFields(TransactionCaseWithUserDemo, TransactionExpressionCase):
         discussion.copy({'messages': [Command.set(message1.ids)]})
         self.assertEqual(message.with_context(lang='fr_FR').label, 'bonjour')
         self.assertFalse(message1.label)
-
-    def test_81_binary_assign(self):
-        binary_value1 = BinaryBytes(b'content')
-        record = self.env['test_orm.model_binary'].create({'binary': binary_value1})
-        self.assertEqual(record.binary.content, binary_value1.content)
-        self.assertEqual(record.binary_related_no_store.content, binary_value1.content)
-        self.assertEqual(record.binary_related_store.content, binary_value1.content)
-
-        binary_value2 = BinaryBytes(b'test')
-        record.binary = binary_value2
-        self.assertEqual(record.binary.content, binary_value2.content)
-        self.assertEqual(record.binary_related_no_store.content, binary_value2.content)
-        self.assertEqual(record.binary_related_store.content, binary_value2.content)
-
-        record.env.invalidate_all()
-        self.assertEqual(record.binary.content, binary_value2.content)
-        self.assertEqual(record.binary_related_no_store.content, binary_value2.content)
-        self.assertEqual(record.binary_related_store.content, binary_value2.content)
-
-    def test_82_binary_bin_size(self):
-        binary_value = BinaryBytes(b'content')
-        record = self.env['test_orm.model_binary'].create({'binary': binary_value})
-
-        def assertBinaryValue(read_value, fields=('binary', 'binary_related_store', 'binary_related_no_store')):
-            for field in fields:
-                self.assertEqual(record[field].content, binary_value.content, f'Incorrect result for {field}')
-            vals = record.read(fields)[0]
-            for field in fields:
-                self.assertEqual(vals[field], read_value)
-
-        record = record.with_context(bin_size=True)
-        assertBinaryValue(f'{binary_value.size}.00 bytes')
-
-        record = record.with_context(bin_size=False)
-        assertBinaryValue(base64.b64encode(binary_value.content).decode())
-
-        # updating and invalidation with bin_size has no effect
-        record = record.with_context(bin_size=True)
-        record.binary = binary_value = BinaryBytes(b'test')
-        record.env.invalidate_all()
-        assertBinaryValue(f'{binary_value.size}.00 bytes')
-
-    def test_85_binary_guess_zip(self):
-        # Regular ZIP files can be uploaded by non-admin users
-        self.env['test_orm.binary_svg'].with_user(self.user_demo).create({
-            'name': 'Test without attachment',
-            'image_wo_attachment': BinaryBytes(ZIP_RAW),
-        })
-
-    def test_86_text_base64_guess_svg(self):
-        with self.assertRaises(UserError) as e:
-            self.env['test_orm.binary_svg'].with_user(self.user_demo).create({
-                'name': 'Test without attachment',
-                'image_wo_attachment': BinaryBytes(SVG_RAW),
-            })
-        self.assertEqual(e.exception.args[0], 'Only admins can upload SVG files.')
-
-    def test_90_binary_svg(self):
-        # This should work without problems
-        self.env['test_orm.binary_svg'].create({
-            'name': 'Test without attachment',
-            'image_wo_attachment': BinaryBytes(SVG_RAW),
-        })
-        # And this gives error
-        with self.assertRaises(UserError):
-            self.env['test_orm.binary_svg'].with_user(
-                self.user_demo,
-            ).create({
-                'name': 'Test without attachment',
-                'image_wo_attachment': BinaryBytes(SVG_RAW),
-            })
-
-    def test_91_binary_svg_attachment(self):
-        # This doesn't neuter SVG with admin
-        record = self.env['test_orm.binary_svg'].create({
-            'name': 'Test without attachment',
-            'image_attachment': BinaryBytes(SVG_RAW),
-        })
-        attachment = self.env['ir.attachment'].search([
-            ('res_model', '=', record._name),
-            ('res_field', '=', 'image_attachment'),
-            ('res_id', '=', record.id),
-        ])
-        self.assertEqual(attachment.mimetype, 'image/svg+xml')
-        # ...but this should be neutered with demo user
-        record = self.env['test_orm.binary_svg'].with_user(
-            self.user_demo,
-        ).create({
-            'name': 'Test without attachment',
-            'image_attachment': BinaryBytes(SVG_RAW),
-        })
-        attachment = self.env['ir.attachment'].search([
-            ('res_model', '=', record._name),
-            ('res_field', '=', 'image_attachment'),
-            ('res_id', '=', record.id),
-        ])
-        self.assertEqual(attachment.mimetype, 'text/plain')
-
-    def test_92_binary_self_avatar_svg(self):
-        demo_user = self.user_demo
-        # User demo changes his own avatar
-        demo_user.with_user(demo_user).image_1920 = BinaryBytes(SVG_RAW)
-        # The SVG file should have been neutered
-        attachment = self.env['ir.attachment'].search([
-            ('res_model', '=', demo_user.partner_id._name),
-            ('res_field', '=', 'image_1920'),
-            ('res_id', '=', demo_user.partner_id.id),
-        ])
-        self.assertEqual(attachment.mimetype, 'text/plain')
-
-    def test_93_monetary_related(self):
-        """ Check the currency field on related monetary fields. """
-        # check base field
-        model = self.env['test_orm.monetary_base']
-        field = model._fields['amount']
-        self.assertEqual(field.get_currency_field(model), 'base_currency_id')
-
-        # related fields must use the field 'currency_id' or 'x_currency_id'
-        model = self.env['test_orm.monetary_related']
-        field = model._fields['amount']
-        self.assertEqual(field.related, 'monetary_id.amount')
-        self.assertEqual(field.get_currency_field(model), 'currency_id')
-
-        model = self.env['test_orm.monetary_custom']
-        field = model._fields['x_amount']
-        self.assertEqual(field.related, 'monetary_id.amount')
-        self.assertEqual(field.get_currency_field(model), 'x_currency_id')
-
-        # inherited field must use the same field as its parent field
-        model = self.env['test_orm.monetary_inherits']
-        field = model._fields['amount']
-        self.assertEqual(field.related, 'monetary_id.amount')
-        self.assertEqual(field.get_currency_field(model), 'base_currency_id')
-
-    def test_94_image(self):
-        f = io.BytesIO()
-        Image.new('RGB', (4000, 2000), '#4169E1').save(f, 'PNG')
-        f.seek(0)
-        image_w = BinaryBytes(f.read())
-
-        f = io.BytesIO()
-        Image.new('RGB', (2000, 4000), '#4169E1').save(f, 'PNG')
-        f.seek(0)
-        image_h = BinaryBytes(f.read())
-
-        def assertResized(record, **sizes):
-            for field_name, size in sizes.items():
-                self.assertEqual(binary_to_image(record[field_name]).size, size, f"Image {field_name} incorrectly sized")
-
-        record = self.env['test_orm.model_image'].create({
-            'name': 'image',
-            'image': image_w,
-            'image_128': image_w,
-        })
-
-        # test create (no resize)
-        self.assertEqual(record.image.content, image_w.content)
-        # test create (resize)
-        # 64: related stored on column
-        # 128: store
-        # 256: related no store
-        # 512: related store
-        assertResized(record, image_64=(64, 32), image_128=(128, 64), image_256=(256, 128), image_512=(512, 256))
-
-        record.write({
-            'image': image_h,
-            'image_128': image_h,
-        })
-
-        # test write (no resize)
-        self.assertEqual(record.image.content, image_h.content)
-        assertResized(record, image_64=(32, 64), image_128=(64, 128), image_256=(128, 256), image_512=(256, 512))
-
-        record = self.env['test_orm.model_image'].create({
-            'name': 'image',
-            'image': image_h,
-            'image_128': image_h,
-        })
-
-        # test create (no resize)
-        self.assertEqual(record.image.content, image_h.content)
-        assertResized(record, image_64=(32, 64), image_128=(64, 128), image_256=(128, 256), image_512=(256, 512))
-
-        record.write({
-            'image': image_w,
-            'image_128': image_w,
-        })
-
-        # test write (no resize)
-        self.assertEqual(record.image.content, image_w.content)
-        assertResized(record, image_64=(64, 32), image_128=(128, 64), image_256=(256, 128), image_512=(512, 256))
-
-        # test create inverse store
-        record = self.env['test_orm.model_image'].create({
-            'name': 'image',
-            'image_512': image_w,
-        })
-        assertResized(record, image_512=(512, 256), image=(4000, 2000), image_256=(256, 128), image_64=(64, 32))
-        # test write inverse store
-        record.write({
-            'image_512': image_h,
-        })
-        assertResized(record, image_512=(256, 512), image=(2000, 4000), image_256=(128, 256), image_64=(32, 64))
-
-        # test create inverse no store
-        record = self.env['test_orm.model_image'].with_context(image_no_postprocess=True).create({
-            'name': 'image',
-            'image_256': image_w,
-        })
-        assertResized(record, image_512=(512, 256), image=(4000, 2000), image_256=(256, 128), image_64=(64, 32))
-        # test write inverse no store
-        record.write({
-            'image_256': image_h,
-        })
-        assertResized(record, image_512=(256, 512), image=(2000, 4000), image_256=(128, 256), image_64=(32, 64))
-
-        # test create inverse stored column
-        record = self.env['test_orm.model_image'].with_context(image_no_postprocess=True).create({
-            'name': 'image',
-            'image_64': image_w,
-        })
-        assertResized(record, image_512=(512, 256), image=(4000, 2000), image_256=(256, 128), image_64=(64, 32))
-        # test write inverse stored column
-        record.write({
-            'image_64': image_h,
-        })
-        assertResized(record, image_512=(256, 512), image=(2000, 4000), image_256=(128, 256), image_64=(32, 64))
-
-        # test bin_size
-        record.invalidate_recordset()
-        values_bin_size = record.with_context(bin_size=True).read([])[0]
-        self.assertEqual(values_bin_size.get('image'), '31.54 Kb')
-        self.assertEqual(values_bin_size.get('image_512'), '1.02 Kb')
-        self.assertEqual(values_bin_size.get('image_256'), '424.00 bytes')
-        self.assertEqual(values_bin_size.get('image_64'), '111.00 bytes')
-
-        # ensure image_data_uri works (value must be bytes and not string)
-        self.assertEqual(record.image_256.to_base64()[:8], 'iVBORw0K')
-        self.assertEqual(image_data_uri(record.image_256)[:30], 'data:image/png;base64,iVBORw0K')
-
-        # ensure invalid image raises
-        with self.assertRaises(UserError):
-            record.write({
-                'image': BinaryBytes(b'invalid image'),
-            })
-
-        # assignment of invalid image on new record does nothing, the value is
-        # taken from origin instead (use-case: onchange)
-        new_record = record.new(origin=record)
-        with mute_logger('odoo.fields'):
-            new_record.image = '31.54 Kb'
-        self.assertEqual(record.image.content, image_h.content)
-        self.assertEqual(new_record.image.content, image_h.content)
-
-        # assignment to new record with origin should not do any query
-        with self.assertQueryCount(0):
-            new_record.image = image_w
-
-    def test_96_order_m2o(self):
-        belgium, congo = self.env['test_orm.country'].create([
-            {'name': "Duchy of Brabant"},
-            {'name': "Congo"},
-        ])
-        cities = self.env['test_orm.city'].create([
-            {'name': "Brussels", 'country_id': belgium.id},
-            {'name': "Kinshasa", 'country_id': congo.id},
-        ])
-        # cities are sorted by country_id, name
-        self.assertEqual(cities.sorted().mapped('name'), ["Kinshasa", "Brussels"])
-
-        # change order of countries, and check sorted()
-        belgium.name = "Belgium"
-        self.assertEqual(cities.sorted().mapped('name'), ["Brussels", "Kinshasa"])
-
-    def test_97_ir_rule_m2m_field(self):
-        """Ensures m2m fields can't be read if the left records can't be read.
-        Also makes sure reading m2m doesn't take more queries than necessary."""
-        tag = self.env['test_orm.multi.tag'].create({})
-        record = self.env['test_orm.multi.line'].create({
-            'name': 'image',
-            'tags': [Command.link(tag.id)],
-        })
-
-        # only one query as admin: reading pivot table
-        with self.assertQueryCount(1):
-            # trick: if value is in cache, read() does not make any query
-            record.invalidate_recordset(['tags'])
-            record.read(['tags'])
-
-        user = self.env['res.users'].create({'name': "user", 'login': "user"})
-        record_user = record.with_user(user)
-
-        # prep the following query count by caching access check related data
-        record_user.invalidate_recordset(['tags'])
-        record_user.read(['tags'])
-
-        # only one query as user: reading pivot table
-        with self.assertQueryCount(1):
-            # trick: if value is in cache, read() does not make any query
-            record_user.invalidate_recordset(['tags'])
-            record_user.read(['tags'])
-
-        # create a passing ir.rule
-        self.env['ir.rule'].create({
-            'model_id': self.env['ir.model']._get(record._name).id,
-            'domain_force': "[('id', '=', %d)]" % record.id,
-        })
-
-        # prep the following query count by caching access check related data
-        record_user.invalidate_recordset(['tags'])
-        record_user.read(['tags'])
-
-        # still only 1 query: reading pivot table
-        # access rules are checked in python in this case
-        with self.assertQueryCount(1):
-            # trick: if value is in cache, read() does not make any query
-            record_user.invalidate_recordset(['tags'])
-            record_user.read(['tags'])
-
-        # create a blocking ir.rule
-        self.env['ir.rule'].create({
-            'model_id': self.env['ir.model']._get(record._name).id,
-            'domain_force': "[('id', '!=', %d)]" % record.id,
-        })
-
-        # ensure ir.rule is applied even when reading m2m
-        with self.assertRaises(AccessError):
-            record_user.read(['tags'])
 
     def test_98_prefetch_translate(self):
         Model = self.registry['test_orm.prefetch']
@@ -3229,376 +2360,9 @@ class TestFields(TransactionCaseWithUserDemo, TransactionExpressionCase):
             for index, record in enumerate(records):
                 record.write({'harry': index + 2})
 
-    def test_html_sanitize(self):
-        record = self.env['test_orm.mixed'].create({})
-        record_value = write_value = "<div>EXTERNAL SUBMISSION - Customer not verified<br>\n<br>\n<p>### TOUR DATA ###</p></div>"
-        record.comment0 = write_value
-        self.assertEqual(record.comment0, record_value)
-        record.invalidate_recordset()
-        self.assertEqual(record.comment0, record_value)
-
     def test_related_column_type(self):
         related_float_field = self.env['test_orm.related']._fields['foo_float_id']
         self.assertEqual(related_float_field.column_type[1], 'numeric')
-
-
-class TestX2many(TransactionExpressionCase):
-
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        cls.user_portal = cls.env['res.users'].sudo().search([('login', '=', 'portal')])
-        cls.partner_portal = cls.user_portal.partner_id
-
-        if not cls.user_portal:
-            cls.env['ir.config_parameter'].sudo().set_int('auth_password_policy.minlength', 4)
-            cls.partner_portal = cls.env['res.partner'].create({
-                'name': 'Joel Willis',
-                'email': 'joel.willis63@example.com',
-            })
-            cls.user_portal = cls.env['res.users'].with_context(no_reset_password=True).create({
-                'login': 'portal',
-                'password': 'portal',
-                'partner_id': cls.partner_portal.id,
-                'group_ids': [Command.set([cls.env.ref('base.group_portal').id])],
-            })
-
-    def test_definition_many2many(self):
-        """ Test the definition of inherited many2many fields. """
-        field = self.env['test_orm.multi.line']._fields['tags']
-        self.assertEqual(field.relation, 'test_orm_multi_line_test_orm_multi_tag_rel')
-        self.assertEqual(field.column1, 'test_orm_multi_line_id')
-        self.assertEqual(field.column2, 'test_orm_multi_tag_id')
-
-        field = self.env['test_orm.multi.line2']._fields['tags']
-        self.assertEqual(field.relation, 'test_orm_multi_line2_test_orm_multi_tag_rel')
-        self.assertEqual(field.column1, 'test_orm_multi_line2_id')
-        self.assertEqual(field.column2, 'test_orm_multi_tag_id')
-
-    def test_10_ondelete_many2many(self):
-        """Test A can't be deleted when used on the relation."""
-        record_a = self.env['test_orm.model_a'].create({'name': 'a'})
-        record_b = self.env['test_orm.model_b'].create({'name': 'b'})
-        record_a.write({
-            'a_restricted_b_ids': [Command.set(record_b.ids)],
-        })
-        with self.assertRaises(psycopg2.IntegrityError):
-            with mute_logger('odoo.sql_db'):
-                record_a.unlink()
-        # Test B is still cascade.
-        record_b.unlink()
-        self.assertFalse(record_b.exists())
-
-    def test_11_ondelete_many2many(self):
-        """Test B can't be deleted when used on the relation."""
-        record_a = self.env['test_orm.model_a'].create({'name': 'a'})
-        record_b = self.env['test_orm.model_b'].create({'name': 'b'})
-        record_a.write({
-            'b_restricted_b_ids': [Command.set(record_b.ids)],
-        })
-        with self.assertRaises(psycopg2.IntegrityError):
-            with mute_logger('odoo.sql_db'):
-                record_b.unlink()
-        # Test A is still cascade.
-        record_a.unlink()
-        self.assertFalse(record_a.exists())
-
-    def test_12_active_test_one2many(self):
-        Model = self.env['test_orm.model_active_field']
-
-        parent = Model.create({})
-        self.assertFalse(parent.children_ids)
-
-        # create with implicit active_test=True in context
-        child1, child2 = Model.create([
-            {'parent_id': parent.id, 'active': True},
-            {'parent_id': parent.id, 'active': False},
-        ])
-        act_children = child1
-        all_children = child1 + child2
-        self.assertEqual(parent.children_ids, act_children)
-        self.assertEqual(parent.with_context(active_test=True).children_ids, act_children)
-        self.assertEqual(parent.with_context(active_test=False).children_ids, all_children)
-
-        # create with active_test=False in context
-        child3, child4 = Model.with_context(active_test=False).create([
-            {'parent_id': parent.id, 'active': True},
-            {'parent_id': parent.id, 'active': False},
-        ])
-        act_children = child1 + child3
-        all_children = child1 + child2 + child3 + child4
-        self.assertEqual(parent.children_ids, act_children)
-        self.assertEqual(parent.with_context(active_test=True).children_ids, act_children)
-        self.assertEqual(parent.with_context(active_test=False).children_ids, all_children)
-
-        # replace active children
-        parent.write({'children_ids': [Command.set([child1.id])]})
-        act_children = child1
-        all_children = child1 + child2 + child4
-        self.assertEqual(parent.children_ids, act_children)
-        self.assertEqual(parent.with_context(active_test=True).children_ids, act_children)
-        self.assertEqual(parent.with_context(active_test=False).children_ids, all_children)
-
-        # replace all children
-        parent.with_context(active_test=False).write({'children_ids': [Command.set([child1.id])]})
-        act_children = child1
-        all_children = child1
-        self.assertEqual(parent.children_ids, act_children)
-        self.assertEqual(parent.with_context(active_test=True).children_ids, act_children)
-        self.assertEqual(parent.with_context(active_test=False).children_ids, all_children)
-
-        # check recomputation of inactive records
-        parent.write({'children_ids': [Command.set(child4.ids)]})
-        self.assertTrue(child4.parent_active)
-        parent.active = False
-        self.assertFalse(child4.parent_active)
-
-    def test_12_active_test_one2many_with_context(self):
-        Model = self.env['test_orm.model_active_field']
-        parent = Model.create({})
-        all_children = Model.create([
-            {'parent_id': parent.id, 'active': True},
-            {'parent_id': parent.id, 'active': False},
-        ])
-        act_children = all_children[0]
-
-        self.assertEqual(parent.children_ids, act_children)
-        self.assertEqual(parent.with_context(active_test=True).children_ids, act_children)
-        self.assertEqual(parent.with_context(active_test=False).children_ids, all_children)
-
-        self.assertEqual(parent.all_children_ids, all_children)
-        self.assertEqual(parent.with_context(active_test=True).all_children_ids, all_children)
-        self.assertEqual(parent.with_context(active_test=False).all_children_ids, all_children)
-
-        self.assertEqual(parent.active_children_ids, act_children)
-        self.assertEqual(parent.with_context(active_test=True).active_children_ids, act_children)
-        self.assertEqual(parent.with_context(active_test=False).active_children_ids, act_children)
-
-        # check read()
-        self.env.invalidate_all()
-        self.assertEqual(parent.children_ids, act_children)
-        self.assertEqual(parent.all_children_ids, all_children)
-        self.assertEqual(parent.active_children_ids, act_children)
-
-        self.env.invalidate_all()
-        self.assertEqual(parent.with_context(active_test=False).children_ids, all_children)
-        self.assertEqual(parent.with_context(active_test=False).all_children_ids, all_children)
-        self.assertEqual(parent.with_context(active_test=False).active_children_ids, act_children)
-
-    def test_12_active_test_one2many_search(self):
-        Model = self.env['test_orm.model_active_field']
-        parent = Model.create({
-            'children_ids': [
-                Command.create({'name': 'A', 'active': True}),
-                Command.create({'name': 'B', 'active': False}),
-            ],
-        })
-
-        # a one2many field without context does not match its inactive children
-        self.assertIn(parent, self._search(Model, [('children_ids.name', '=', 'A')]))
-        self.assertNotIn(parent, self._search(Model, [('children_ids.name', '=', 'B')]))
-        # Same result when it used _search_display_name
-        self.assertIn(parent, self._search(Model, [('children_ids', '=', 'A')]))
-        self.assertNotIn(parent, self._search(Model, [('children_ids', '=', 'B')]))
-        # Same result with the child_of operator
-        self.assertIn(parent, self._search(Model, [('children_ids', 'child_of', 'A')]))
-        self.assertNotIn(parent, self._search(Model, [('children_ids', 'child_of', 'B')]))
-
-        # a one2many field with active_test=False matches its inactive children
-        self.assertIn(parent, self._search(Model, [('all_children_ids.name', '=', 'A')]))
-        self.assertIn(parent, self._search(Model, [('all_children_ids.name', '=', 'B')]))
-        # Same result when it used _search_display_name
-        self.assertIn(parent, self._search(Model, [('all_children_ids', '=', 'A')]))
-        # Same result with the child_of operator
-        self.assertIn(parent, self._search(Model, [('all_children_ids', 'child_of', 'A')]))
-        self.assertIn(parent, self._search(Model, [('all_children_ids', '=', 'B')]))
-        # Same result with the child_of operator
-        self.assertIn(parent, self._search(Model, [('all_children_ids', 'child_of', 'A')]))
-        self.assertIn(parent, self._search(Model, [('all_children_ids', 'child_of', 'B')]))
-
-    def test_12_active_test_many2many_search(self):
-        Model = self.env['test_orm.model_active_field']
-        parent = Model.create({
-            'relatives_ids': [
-                Command.create({'name': 'A', 'active': True}),
-                Command.create({'name': 'B', 'active': False}),
-            ],
-        })
-        child_a, child_b = parent.with_context(active_test=False).relatives_ids
-        # TODO all_relatives_ids is empty, because it is another fields using
-        # the same backend table as relative_ids
-        Model.invalidate_model(['all_relatives_ids'])
-
-        # a many2many field without context does not match its inactive children
-        self.assertIn(parent, self._search(Model, [('relatives_ids.name', '=', 'A')]))
-        self.assertNotIn(parent, self._search(Model, [('relatives_ids.name', '=', 'B')]))
-        # Same result when it used _search_display_name
-        self.assertIn(parent, self._search(Model, [('relatives_ids', '=', 'A')]))
-        self.assertNotIn(parent, self._search(Model, [('relatives_ids', '=', 'B')]))
-        # Same result with the child_of operator
-        self.assertIn(parent, self._search(Model, [('relatives_ids', 'child_of', child_a.id)]))
-        self.assertIn(parent, self._search(Model, [('relatives_ids', 'child_of', 'A')]))
-        self.assertNotIn(parent, self._search(Model, [('relatives_ids', 'child_of', child_b.id)]))
-        self.assertNotIn(parent, self._search(Model, [('relatives_ids', 'child_of', 'B')]))
-
-        # a many2many field with active_test=False matches its inactive children
-        self.assertIn(parent, self._search(Model, [('all_relatives_ids.name', '=', 'A')]))
-        self.assertIn(parent, self._search(Model, [('all_relatives_ids.name', '=', 'B')]))
-        # Same result when it used _search_display_name
-        self.assertIn(parent, self._search(Model, [('all_relatives_ids', '=', 'A')]))
-        self.assertIn(parent, self._search(Model, [('all_relatives_ids', '=', 'B')]))
-        # Same result with the child_of operator
-        self.assertIn(parent, self._search(Model, [('all_relatives_ids', 'child_of', child_a.id)]))
-        self.assertIn(parent, self._search(Model, [('all_relatives_ids', 'child_of', 'A')]))
-        self.assertIn(parent, self._search(Model, [('all_relatives_ids', 'child_of', child_b.id)]))
-        self.assertIn(parent, self._search(Model, [('all_relatives_ids', 'child_of', 'B')]))
-
-    def test_search_many2many(self):
-        """ Tests search on many2many fields. """
-        tags = self.env['test_orm.multi.tag']
-        tagA = tags.create({})
-        tagB = tags.create({})
-        tagC = tags.create({})
-        recs = self.env['test_orm.multi.line']
-        recW = recs.create({})
-        recX = recs.create({'tags': [Command.link(tagA.id)]})
-        recY = recs.create({'tags': [Command.link(tagB.id)]})
-        recZ = recs.create({'tags': [Command.link(tagA.id), Command.link(tagB.id)]})
-        recs = recW + recX + recY + recZ
-
-        # test 'in'
-        result = self._search(recs, [('tags', 'in', (tagA + tagB).ids)])
-        self.assertEqual(result, recX + recY + recZ)
-
-        result = self._search(recs, [('tags', 'in', tagA.ids)])
-        self.assertEqual(result, recX + recZ)
-
-        result = self._search(recs, [('tags', 'in', tagB.ids)])
-        self.assertEqual(result, recY + recZ)
-
-        result = self._search(recs, [('tags', 'in', tagC.ids)])
-        self.assertEqual(result, recs.browse())
-
-        result = self._search(recs, [('tags', 'in', [])])
-        self.assertEqual(result, recs.browse())
-
-        # test 'not in'
-        result = self._search(recs, [('id', 'in', recs.ids), ('tags', 'not in', (tagA + tagB).ids)])
-        self.assertEqual(result, recs - recX - recY - recZ)
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('tags', 'not in', tagA.ids)])
-        self.assertEqual(result, recs - recX - recZ)
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('tags', 'not in', tagB.ids)])
-        self.assertEqual(result, recs - recY - recZ)
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('tags', 'not in', tagC.ids)])
-        self.assertEqual(result, recs)
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('tags', 'not in', [])])
-        self.assertEqual(result, recs)
-
-        # special case: compare with False
-        result = self._search(recs, [('id', 'in', recs.ids), ('tags', '=', False)])
-        self.assertEqual(result, recW)
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('tags', '!=', False)])
-        self.assertEqual(result, recs - recW)
-
-    def test_search_one2many(self):
-        """ Tests search on one2many fields. """
-        recs = self.env['test_orm.multi']
-        recX = recs.create({'lines': [Command.create({}), Command.create({})]})
-        recY = recs.create({'lines': [Command.create({})]})
-        recZ = recs.create({})
-        recs = recX + recY + recZ
-        line1, line2, line3 = recs.lines
-        line4 = recs.create({'lines': [Command.create({})]}).lines
-        line0 = line4.create({})
-
-        # test 'in'
-        result = self._search(recs, [('id', 'in', recs.ids), ('lines', 'in', (line1 + line2 + line3 + line4).ids)])
-        self.assertEqual(result, recX + recY)
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('lines', 'in', (line1 + line3 + line4).ids)])
-        self.assertEqual(result, recX + recY)
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('lines', 'in', (line1 + line4).ids)])
-        self.assertEqual(result, recX)
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('lines', 'in', line4.ids)])
-        self.assertEqual(result, recs.browse())
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('lines', 'in', [])])
-        self.assertEqual(result, recs.browse())
-
-        # test 'not in'
-        result = self._search(recs, [('id', 'in', recs.ids), ('lines', 'not in', (line1 + line2 + line3).ids)])
-        self.assertEqual(result, recs - recX - recY)
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('lines', 'not in', (line1 + line3).ids)])
-        self.assertEqual(result, recs - recX - recY)
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('lines', 'not in', line1.ids)])
-        self.assertEqual(result, recs - recX)
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('lines', 'not in', (line1 + line4).ids)])
-        self.assertEqual(result, recs - recX)
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('lines', 'not in', line4.ids)])
-        self.assertEqual(result, recs)
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('lines', 'not in', [])])
-        self.assertEqual(result, recs)
-
-        # test 'not in' where the lines contain NULL values
-        result = self._search(recs, [('id', 'in', recs.ids), ('lines', 'not in', (line1 + line0).ids)])
-        self.assertEqual(result, recs - recX)
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('lines', 'not in', line0.ids)])
-        self.assertEqual(result, recs)
-
-        # special case: compare with False
-        result = self._search(recs, [('id', 'in', recs.ids), ('lines', '=', False)])
-        self.assertEqual(result, recZ)
-
-        result = self._search(recs, [('id', 'in', recs.ids), ('lines', '!=', False)])
-        self.assertEqual(result, recs - recZ)
-
-    def test_create_batch_m2m(self):
-        lines = self.env['test_orm.multi.line'].create([{
-            'tags': [Command.create({'name': str(j)}) for j in range(3)],
-        } for i in range(3)])
-        self.assertEqual(len(lines), 3)
-        for line in lines:
-            self.assertEqual(len(line.tags), 3)
-
-    def test_custom_m2m(self):
-        model_id = self.env['ir.model']._get_id('res.partner')
-        field = self.env['ir.model.fields'].create({
-            'name': 'x_foo',
-            'field_description': 'Foo',
-            'model_id': model_id,
-            'ttype': 'many2many',
-            'relation': 'res.country',
-            'store': False,
-        })
-        self.assertTrue(field.unlink())
-
-    def test_custom_m2m_related(self):
-        # this checks the ondelete of a related many2many field
-        model_id = self.env['ir.model']._get_id('res.partner')
-        field = self.env['ir.model.fields'].create({
-            'name': 'x_foo',
-            'field_description': 'Foo',
-            'model_id': model_id,
-            'ttype': 'many2many',
-            'relation': 'res.partner.category',
-            'related': 'category_id',
-            'readonly': True,
-            'store': True,
-        })
-        self.assertTrue(field.unlink())
 
 
 class SudoCommands(TransactionCaseWithUserDemo):
@@ -3733,169 +2497,6 @@ class SudoCommands(TransactionCaseWithUserDemo):
                 my_user.write({
                     'group_ids': [Command.delete(my_user.group_ids[0].id)],
                 })
-
-
-@tagged('at_install', '-post_install')  # LEGACY at_install
-class TestHtmlField(TransactionCase):
-
-    def setUp(self):
-        super().setUp()
-        self.model = self.env['test_orm.mixed']
-
-    def test_00_sanitize(self):
-        self.assertEqual(self.model._fields['comment1'].sanitize, False)
-        self.assertEqual(self.model._fields['comment2'].sanitize_attributes, True)
-        self.assertEqual(self.model._fields['comment2'].strip_classes, False)
-        self.assertEqual(self.model._fields['comment3'].sanitize_attributes, True)
-        self.assertEqual(self.model._fields['comment3'].strip_classes, True)
-
-        some_ugly_html = """<p>Oops this should maybe be sanitized
-% if object.some_field and not object.oriented:
-<table>
-    % if object.other_field:
-    <tr style="margin: 0px; border: 10px solid black;">
-        ${object.mako_thing}
-        <td>
-    </tr>
-    <tr class="custom_class">
-        This is some html.
-    </tr>
-    % endif
-    <tr>
-%if object.dummy_field:
-        <p>Youpie</p>
-%endif"""
-
-        record = self.model.create({
-            'comment1': some_ugly_html,
-            'comment2': some_ugly_html,
-            'comment3': some_ugly_html,
-            'comment4': some_ugly_html,
-        })
-
-        self.assertEqual(record.comment1, some_ugly_html, 'Error in HTML field: content was sanitized but field has sanitize=False')
-
-        self.assertIn('<tr class="', record.comment2)
-
-        # sanitize should have closed tags left open in the original html
-        self.assertIn('</table>', record.comment3, 'Error in HTML field: content does not seem to have been sanitized despise sanitize=True')
-        self.assertIn('</td>', record.comment3, 'Error in HTML field: content does not seem to have been sanitized despise sanitize=True')
-        self.assertIn('<tr style="', record.comment3, 'Style attr should not have been stripped')
-        # sanitize does not keep classes if asked to
-        self.assertNotIn('<tr class="', record.comment3)
-
-        self.assertNotIn('<tr style="', record.comment4, 'Style attr should have been stripped')
-
-    def test_01_sanitize_groups(self):
-        self.assertEqual(self.model._fields['comment5'].sanitize, True)
-        self.assertEqual(self.model._fields['comment5'].sanitize_overridable, True)
-
-        internal_user = self.env['res.users'].create({
-            'name': 'test internal user',
-            'login': 'test_sanitize',
-            'group_ids': [(6, 0, [self.ref('base.group_user')])],
-        })
-        bypass_user = self.env['res.users'].create({
-            'name': 'test bypass user',
-            'login': 'test_sanitize2',
-            'group_ids': [(6, 0, [self.ref('base.group_user'), self.ref('base.group_sanitize_override')])],
-        })
-        record = self.env['test_orm.mixed'].create({})
-
-        # 1. Test normalize case: diff due to normalize should not prevent the
-        #    changes
-        val = '<blockquote>Something</blockquote>'
-        normalized_val = '<blockquote data-o-mail-quote-node="1" data-o-mail-quote="1">Something</blockquote>'
-        write_vals = {'comment5': val}
-
-        record.with_user(internal_user).write(write_vals)
-        self.assertEqual(record.comment5, normalized_val,
-                         "should be normalized (not in groups)")
-        record.with_user(bypass_user).write(write_vals)
-        self.assertEqual(record.comment5, val,
-                         "should not be normalized (has group)")
-        record.with_user(internal_user).write(write_vals)
-        self.assertEqual(record.comment5, normalized_val,
-                         "should be normalized (not in groups) despite admin previous diff")
-
-        # 2. Test main use case: prevent restricted user to wipe non restricted
-        #    user previous change
-        val = '<script></script>'
-        write_vals = {'comment5': val}
-
-        record.with_user(internal_user).write(write_vals)
-        self.assertEqual(record.comment5, '',
-                         "should be sanitized (not in groups)")
-        record.with_user(bypass_user).write(write_vals)
-        self.assertEqual(record.comment5, val,
-                         "should not be sanitized (has group)")
-        with self.assertRaises(UserError):
-            # should crash (not in groups and sanitize would break content of
-            # other user that bypassed the sanitize)
-            record.with_user(internal_user).write(write_vals)
-
-        # 3. Make sure field compare in `_convert` is working as expected with
-        #    special content / format
-        val = '<span  attr1 ="att1"   attr2=\'attr2\'>é@&nbsp;</span><p><span/></p>'
-        write_vals = {'comment5': val}
-        # Once sent through `html_sanitize()` this is becoming:
-        # `<span attr1="att1" attr2="attr2">é@\xa0</span><p><span></span></p>`
-        # Notice those change:
-        # -     `attr1 =` -> `attr1=`    (space before `=`)
-        # -    `   attr2` -> ` attr2`    (multi space -> single space)
-        # -  `=\'attr2\'` -> `="attr2"`  (escaped single quote -> double quote)
-        # -      `&nbsp;` -> `\xa0`
-        # Still, those 2 archs should be considered equals and not raise
-
-        record.with_user(bypass_user).write(write_vals)
-        # Next write shouldn't raise a sanitize right error
-        record.with_user(internal_user).write(write_vals)
-
-        # 4. Ensure our exception handling is fine
-        val = '<!-- I am a comment -->'
-        write_vals = {'comment5': val}
-        record.with_user(internal_user).write(write_vals)
-        self.assertEqual(record.comment5, '',
-                         "should be sanitized (not in groups)")
-
-        # extra test with new record having 'record' as origin
-        new_record = record.new(origin=record)
-        new_record.with_user(bypass_user).comment5
-
-        # this was causing an infinite recursion (see explanation in fields.py)
-        new_record.invalidate_recordset()
-        new_record.with_user(internal_user).comment5
-
-    @patch('odoo.orm.fields_textual.html_sanitize', return_value='<p>comment</p>')
-    def test_onchange_sanitize(self, patch):
-        self.assertTrue(self.registry['test_orm.mixed'].comment2.sanitize)
-
-        record = self.env['test_orm.mixed'].create({
-            'comment2': '<p>comment</p>',
-        })
-
-        # the new value is sanitized upon insertion in db,
-        # but not put in cache, therefore not sanitized a second time
-        self.assertEqual(patch.call_count, 1)
-
-        # new value sanitized for insertion in cache
-        record.comment2 = '<p>comment</p>'
-        self.assertEqual(patch.call_count, 2)
-
-        # the value in cache is dirty -> convert_to_column_update(..., validate=False),
-        # so no additional call to `html_sanitize`
-        record.flush_recordset()
-        self.assertEqual(patch.call_count, 2)
-
-        # value coming from db does not need to be sanitized
-        record.invalidate_recordset()
-        record.comment2
-        self.assertEqual(patch.call_count, 2)
-
-        # value coming from db during an onchange does not need to be sanitized
-        new_record = record.new(origin=record)
-        new_record.comment2
-        self.assertEqual(patch.call_count, 2)
 
 
 @tagged('at_install', '-post_install')  # LEGACY at_install
@@ -4160,435 +2761,6 @@ class TestParentStore(TransactionCaseWithUserDemo):
             Dis.with_user(self.user_demo).search([('categories', 'parent_of', self.cats(3).ids)]),
             dis_0 + dis_3,
         )
-
-
-@tagged('at_install', '-post_install')  # LEGACY at_install
-class TestRequiredMany2one(TransactionCase):
-
-    def test_explicit_ondelete(self):
-        field = self.env['test_orm.req_m2o']._fields['foo']
-        self.assertEqual(field.ondelete, 'cascade')
-
-    def test_implicit_ondelete(self):
-        field = self.env['test_orm.req_m2o']._fields['bar']
-        self.assertEqual(field.ondelete, 'restrict')
-
-    def test_explicit_set_null(self):
-        Model = self.env['test_orm.req_m2o']
-        field = Model._fields['foo']
-
-        # clean up registry after this test
-        self.addCleanup(self.registry.reset_changes)
-        self.patch(field, 'ondelete', 'set null')
-
-        with self.assertRaises(ValueError):
-            field.setup_nonrelated(Model)
-
-
-@tagged('at_install', '-post_install')  # LEGACY at_install
-class TestRequiredMany2oneTransient(TransactionCase):
-
-    def test_explicit_ondelete(self):
-        field = self.env['test_orm.req_m2o_transient']._fields['foo']
-        self.assertEqual(field.ondelete, 'restrict')
-
-    def test_implicit_ondelete(self):
-        field = self.env['test_orm.req_m2o_transient']._fields['bar']
-        self.assertEqual(field.ondelete, 'cascade')
-
-    def test_explicit_set_null(self):
-        Model = self.env['test_orm.req_m2o_transient']
-        field = Model._fields['foo']
-
-        # clean up registry after this test
-        self.addCleanup(self.registry.reset_changes)
-        self.patch(field, 'ondelete', 'set null')
-
-        with self.assertRaises(ValueError):
-            field.setup_nonrelated(Model)
-
-
-@tagged('m2oref')
-class TestMany2oneReference(TransactionExpressionCase):
-
-    def test_delete_m2o_reference_records(self):
-        m = self.env['test_orm.model_many2one_reference']
-        self.env.cr.execute("SELECT max(id) FROM test_orm_model_many2one_reference")
-        ids = self.env.cr.fetchone()
-        # fake record to emulate the unlink of a non-existant record
-        foo = m.browse(1 if not ids[0] else (ids[0] + 1))
-        self.assertTrue(foo.unlink())
-
-    def test_search_inverse_one2many_bypass_search_access(self):
-        record = self.env['test_orm.inverse_m2o_ref'].create({})
-
-        # the one2many field 'model_ids' should be bypass_search_access=True
-        self.patch(type(record).model_ids, 'bypass_search_access', True)
-
-        # create a reference to record
-        reference = self.env['test_orm.model_many2one_reference'].create({'res_id': record.id})
-        reference.res_model = record._name
-
-        # the model field 'res_model' is not in database yet
-        self.assertIn(reference.id, self.env._field_dirty[reference._fields['res_model']])
-
-        # searching on the one2many should flush the field 'res_model'
-        records = record.search([('model_ids.create_date', '!=', False)])
-        self.assertIn(record, records)
-
-        # filtered should be aligned
-        # TODO right now, need to invalidate because the inverse of
-        # many2one_reference is not updated
-        record.invalidate_model()
-        self._search(record, [('model_ids.create_date', '!=', False)])
-
-
-@tagged('selection_abstract')
-@tagged('at_install', '-post_install')  # LEGACY at_install
-class TestSelectionDeleteUpdate(TransactionCase):
-
-    MODEL_ABSTRACT = 'test_orm.state_mixin'
-
-    def setUp(self):
-        super().setUp()
-        # enable unlinking ir.model.fields.selection
-        self.patch(self.registry, 'ready', False)
-
-    def test_unlink_asbtract(self):
-        self.env['ir.model.fields.selection'].search([
-            ('field_id.model', '=', self.MODEL_ABSTRACT),
-            ('field_id.name', '=', 'state'),
-            ('value', '=', 'confirmed'),
-        ], limit=1).unlink()
-
-
-@tagged('selection_update_base')
-@tagged('at_install', '-post_install')  # LEGACY at_install
-class TestSelectionUpdates(TransactionCase):
-    MODEL_BASE = 'test_orm.model_selection_base'
-    MODEL_RELATED = 'test_orm.model_selection_related'
-    MODEL_RELATED_UPDATE = 'test_orm.model_selection_related_updatable'
-
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        # Specifying a lang in env/context should not increase query counts
-        # of CRUD operations
-        cls.env = cls.env(context={'lang': 'en_US'})
-
-    def test_selection(self):
-        self.env[self.MODEL_BASE].create({})   # warming up
-        with self.assertQueryCount(1):
-            self.env[self.MODEL_BASE].create({})
-        with self.assertQueryCount(1):
-            record = self.env[self.MODEL_BASE].create({'my_selection': 'foo'})
-        with self.assertQueryCount(1):
-            record.my_selection = 'bar'
-
-    def test_selection_related_readonly(self):
-        related_record = self.env[self.MODEL_BASE].create({'my_selection': 'foo'})
-        with self.assertQueryCount(2):  # defaults (readonly related field), INSERT
-            record = self.env[self.MODEL_RELATED].create({'selection_id': related_record.id})
-        with self.assertQueryCount(0):
-            record.related_selection = 'bar'
-
-    def test_selection_related(self):
-        related_record = self.env[self.MODEL_BASE].create({'my_selection': 'foo'})
-        with self.assertQueryCount(2):  # defaults (related field), INSERT
-            record = self.env[self.MODEL_RELATED_UPDATE].create({'selection_id': related_record.id})
-        with self.assertQueryCount(2):
-            record.related_selection = 'bar'
-
-
-@tagged('selection_ondelete_base')
-@tagged('at_install', '-post_install')  # LEGACY at_install
-class TestSelectionOndelete(TransactionCase):
-
-    MODEL_BASE = 'test_orm.model_selection_base'
-    MODEL_REQUIRED = 'test_orm.model_selection_required'
-    MODEL_NONSTORED = 'test_orm.model_selection_non_stored'
-    MODEL_WRITE_OVERRIDE = 'test_orm.model_selection_required_for_write_override'
-    MODEL_COMPANY_DEPENDENT = 'test_orm.model_selection_company_dependent'
-
-    def setUp(self):
-        super().setUp()
-        # enable unlinking ir.model.fields.selection
-        self.patch(self.registry, 'ready', False)
-
-    def _unlink_option(self, model, option):
-        self.env['ir.model.fields.selection'].search([
-            ('field_id.model', '=', model),
-            ('field_id.name', '=', 'my_selection'),
-            ('value', '=', option),
-        ], limit=1).unlink()
-
-    def test_ondelete_default(self):
-        # create some records, one of which having the extended selection option
-        rec1 = self.env[self.MODEL_REQUIRED].create({'my_selection': 'foo'})
-        rec2 = self.env[self.MODEL_REQUIRED].create({'my_selection': 'bar'})
-        rec3 = self.env[self.MODEL_REQUIRED].create({'my_selection': 'baz'})
-
-        # test that all values are correct before the removal of the value
-        self.assertEqual(rec1.my_selection, 'foo')
-        self.assertEqual(rec2.my_selection, 'bar')
-        self.assertEqual(rec3.my_selection, 'baz')
-
-        # unlink the extended option (simulates a module uninstall)
-        self._unlink_option(self.MODEL_REQUIRED, 'baz')
-
-        # verify that the ondelete policy has succesfully been applied
-        self.assertEqual(rec1.my_selection, 'foo')
-        self.assertEqual(rec2.my_selection, 'bar')
-        self.assertEqual(rec3.my_selection, 'foo')   # reset to default
-
-    def test_ondelete_base_null_explicit(self):
-        rec1 = self.env[self.MODEL_BASE].create({'my_selection': 'foo'})
-        rec2 = self.env[self.MODEL_BASE].create({'my_selection': 'bar'})
-        rec3 = self.env[self.MODEL_BASE].create({'my_selection': 'quux'})
-
-        self.assertEqual(rec1.my_selection, 'foo')
-        self.assertEqual(rec2.my_selection, 'bar')
-        self.assertEqual(rec3.my_selection, 'quux')
-
-        self._unlink_option(self.MODEL_BASE, 'quux')
-
-        self.assertEqual(rec1.my_selection, 'foo')
-        self.assertEqual(rec2.my_selection, 'bar')
-        self.assertFalse(rec3.my_selection)
-
-    def test_ondelete_base_null_implicit(self):
-        rec1 = self.env[self.MODEL_BASE].create({'my_selection': 'foo'})
-        rec2 = self.env[self.MODEL_BASE].create({'my_selection': 'bar'})
-        rec3 = self.env[self.MODEL_BASE].create({'my_selection': 'ham'})
-
-        self.assertEqual(rec1.my_selection, 'foo')
-        self.assertEqual(rec2.my_selection, 'bar')
-        self.assertEqual(rec3.my_selection, 'ham')
-
-        self._unlink_option(self.MODEL_BASE, 'ham')
-
-        self.assertEqual(rec1.my_selection, 'foo')
-        self.assertEqual(rec2.my_selection, 'bar')
-        self.assertFalse(rec3.my_selection)
-
-    def test_ondelete_cascade(self):
-        rec1 = self.env[self.MODEL_REQUIRED].create({'my_selection': 'foo'})
-        rec2 = self.env[self.MODEL_REQUIRED].create({'my_selection': 'bar'})
-        rec3 = self.env[self.MODEL_REQUIRED].create({'my_selection': 'eggs'})
-
-        self.assertEqual(rec1.my_selection, 'foo')
-        self.assertEqual(rec2.my_selection, 'bar')
-        self.assertEqual(rec3.my_selection, 'eggs')
-
-        self._unlink_option(self.MODEL_REQUIRED, 'eggs')
-
-        self.assertEqual(rec1.my_selection, 'foo')
-        self.assertEqual(rec2.my_selection, 'bar')
-        self.assertFalse(rec3.exists())
-
-    def test_ondelete_literal(self):
-        rec1 = self.env[self.MODEL_REQUIRED].create({'my_selection': 'foo'})
-        rec2 = self.env[self.MODEL_REQUIRED].create({'my_selection': 'bar'})
-        rec3 = self.env[self.MODEL_REQUIRED].create({'my_selection': 'bacon'})
-
-        self.assertEqual(rec1.my_selection, 'foo')
-        self.assertEqual(rec2.my_selection, 'bar')
-        self.assertEqual(rec3.my_selection, 'bacon')
-
-        self._unlink_option(self.MODEL_REQUIRED, 'bacon')
-
-        self.assertEqual(rec1.my_selection, 'foo')
-        self.assertEqual(rec2.my_selection, 'bar')
-        self.assertEqual(rec3.my_selection, 'bar')
-
-    def test_ondelete_multiple_explicit(self):
-        rec1 = self.env[self.MODEL_REQUIRED].create({'my_selection': 'foo'})
-        rec2 = self.env[self.MODEL_REQUIRED].create({'my_selection': 'eevee'})
-        rec3 = self.env[self.MODEL_REQUIRED].create({'my_selection': 'pikachu'})
-
-        self.assertEqual(rec1.my_selection, 'foo')
-        self.assertEqual(rec2.my_selection, 'eevee')
-        self.assertEqual(rec3.my_selection, 'pikachu')
-
-        self._unlink_option(self.MODEL_REQUIRED, 'eevee')
-        self._unlink_option(self.MODEL_REQUIRED, 'pikachu')
-
-        self.assertEqual(rec1.my_selection, 'foo')
-        self.assertEqual(rec2.my_selection, 'bar')
-        self.assertEqual(rec3.my_selection, 'foo')
-
-    def test_ondelete_callback(self):
-        rec = self.env[self.MODEL_REQUIRED].create({'my_selection': 'knickers'})
-
-        self.assertEqual(rec.my_selection, 'knickers')
-
-        self._unlink_option(self.MODEL_REQUIRED, 'knickers')
-
-        self.assertEqual(rec.my_selection, 'foo')
-        self.assertFalse(rec.active)
-
-    def test_non_stored_selection(self):
-        rec = self.env[self.MODEL_NONSTORED].create({})
-        rec.my_selection = 'foo'
-
-        self.assertEqual(rec.my_selection, 'foo')
-
-        self._unlink_option(self.MODEL_NONSTORED, 'foo')
-
-        self.assertFalse(rec.my_selection)
-
-    def test_required_base_selection_field(self):
-        # test that no ondelete action is executed on a required selection field that is not
-        # extended, only required fields that extend it with selection_add should
-        # have ondelete actions defined
-        rec = self.env[self.MODEL_REQUIRED].create({'my_selection': 'foo'})
-        self.assertEqual(rec.my_selection, 'foo')
-
-        self._unlink_option(self.MODEL_REQUIRED, 'foo')
-        self.assertEqual(rec.my_selection, 'foo')
-
-    @mute_logger('odoo.addons.base.models.ir_model')
-    def test_write_override_selection(self):
-        # test that on override to write that raises an error does not prevent the ondelete
-        # policy from executing and cleaning up what needs to be cleaned up
-        rec = self.env[self.MODEL_WRITE_OVERRIDE].create({'my_selection': 'divinity'})
-        self.assertEqual(rec.my_selection, 'divinity')
-
-        self._unlink_option(self.MODEL_WRITE_OVERRIDE, 'divinity')
-        self.assertEqual(rec.my_selection, 'foo')
-
-    def test_ondelete_company_dependent_null_implicit_with_multicompany(self):
-        Model = self.env[self.MODEL_COMPANY_DEPENDENT]
-        company_2 = self.env['res.company'].create({'name': 'Test Company'})
-
-        # create records with the extended selection option
-        records = r1, r2, r3 = Model.create([
-            {'my_selection': 'manual'},
-            {'my_selection': 'auto'},
-            {'my_selection': 'semi_auto'},
-        ])
-
-        # set different values for company_2
-        r1.with_company(company_2).write({'my_selection': 'semi_auto'})
-        r2.with_company(company_2).write({'my_selection': 'semi_auto'})
-        r3.with_company(company_2).write({'my_selection': 'manual'})
-
-        # sanity checks before unlink
-        self.assertEqual(records.mapped("my_selection"), ["manual", "auto", "semi_auto"])
-        self.assertEqual(
-            records.with_company(company_2).mapped("my_selection"),
-            ["semi_auto", "semi_auto", "manual"],
-        )
-
-        # simulates a module uninstall
-        self._unlink_option(self.MODEL_COMPANY_DEPENDENT, 'semi_auto')
-
-        # test that values are removed from all the companies
-        self.assertEqual(records.mapped("my_selection"), ["manual", "auto", False])
-        self.assertEqual(
-            records.with_company(company_2).mapped("my_selection"),
-            [False, False, "manual"],
-        )
-
-
-@tagged('selection_ondelete_advanced')
-@tagged('at_install', '-post_install')  # LEGACY at_install
-class TestSelectionOndeleteAdvanced(TransactionCase):
-
-    MODEL_BASE = 'test_orm.model_selection_base'
-    MODEL_REQUIRED = 'test_orm.model_selection_required'
-
-    def setUp(self):
-        super().setUp()
-        # necessary cleanup for resetting changes in the registry
-        for model_name in (self.MODEL_BASE, self.MODEL_REQUIRED):
-            Model = self.registry[model_name]
-            self.addCleanup(setattr, Model, '_base_classes__', Model._base_classes__)
-
-    def test_ondelete_unexisting_policy(self):
-        from odoo.orm.model_classes import add_to_registry  # noqa: PLC0415
-
-        class Foo(models.Model):
-            _module = None
-            _name = self.MODEL_REQUIRED
-            _inherit = [self.MODEL_REQUIRED]
-
-            my_selection = fields.Selection(selection_add=[
-                ('random', "Random stuff"),
-            ], ondelete={'random': 'poop'})
-
-        add_to_registry(self.registry, Foo)
-
-        with self.assertRaises(ValueError):
-            self.registry._setup_models__(self.env.cr, [])  # incremental setup
-
-    def test_ondelete_default_no_default(self):
-        from odoo.orm.model_classes import add_to_registry  # noqa: PLC0415
-
-        class Foo(models.Model):
-            _module = None
-            _name = self.MODEL_BASE
-            _inherit = [self.MODEL_BASE]
-
-            my_selection = fields.Selection(selection_add=[
-                ('corona', "Corona beers suck"),
-            ], ondelete={'corona': 'set default'})
-
-        add_to_registry(self.registry, Foo)
-
-        with self.assertRaises(AssertionError):
-            self.registry._setup_models__(self.env.cr, [])  # incremental setup
-
-    def test_ondelete_value_no_valid(self):
-        from odoo.orm.model_classes import add_to_registry  # noqa: PLC0415
-
-        class Foo(models.Model):
-            _module = None
-            _name = self.MODEL_BASE
-            _inherit = [self.MODEL_BASE]
-
-            my_selection = fields.Selection(selection_add=[
-                ('westvleteren', "Westvleteren beers is overrated"),
-            ], ondelete={'westvleteren': 'set foooo'})
-
-        add_to_registry(self.registry, Foo)
-
-        with self.assertRaises(AssertionError):
-            self.registry._setup_models__(self.env.cr, [])  # incremental setup
-
-    def test_ondelete_required_null_explicit(self):
-        from odoo.orm.model_classes import add_to_registry  # noqa: PLC0415
-
-        class Foo(models.Model):
-            _module = None
-            _name = self.MODEL_REQUIRED
-            _inherit = [self.MODEL_REQUIRED]
-
-            my_selection = fields.Selection(selection_add=[
-                ('brap', "Brap"),
-            ], ondelete={'brap': 'set null'})
-
-        add_to_registry(self.registry, Foo)
-
-        with self.assertRaises(ValueError):
-            self.registry._setup_models__(self.env.cr, [])  # incremental setup
-
-    def test_ondelete_required_null_implicit(self):
-        from odoo.orm.model_classes import add_to_registry  # noqa: PLC0415
-
-        class Foo(models.Model):
-            _module = None
-            _name = self.MODEL_REQUIRED
-            _inherit = [self.MODEL_REQUIRED]
-
-            my_selection = fields.Selection(selection_add=[
-                ('boing', "Boyoyoyoing"),
-            ])
-
-        add_to_registry(self.registry, Foo)
-
-        with self.assertRaises(ValueError):
-            self.registry._setup_models__(self.env.cr, [])  # incremental setup
 
 
 @tagged('at_install', '-post_install')  # LEGACY at_install
